requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
numpy==1.23.4
Pillow==9.2.0
pip==22.2.2
python-dotenv==0.21.0
requests==2.28.1
setuptools==65.3.0
urllib3==1.26.12
wheel==0.37.1
yarl==1.8.1